refactor: Move EnvironmentSelect
